requirements.txt
flask==3.1.1
tensorflow==2.18.0
numpy==2.0.2
Pillow==11.1.0
joblib==1.4.2
scikit-learn==1.6.1
gunicorn==23.0.0
streamlit==1.43.2
